moved title back into MainTitle component in the NavBar
