+++ src/components/NavBar/MainTitle.styles.ts
@@ -1,46 +1,62 @@
 import { CSSObject } from '@emotion/react';
-import { DynamicStyles } from '../../types/StyleTypes';
 
-export const headlineStyle: CSSObject = {
-	fontSize: '3rem',
-	lineHeight: '1',
-	margin: '0',
-	boxSizing: 'border-box',
-	paddingBottom: '1rem',
+export const hexagonBoxStyle: CSSObject = {
+	width: '100vmin',
+	height: '100vmin',
+	display: 'grid',
+	gridTemplateColumns: '1fr',
+	gridTemplateRows: '1fr',
+	zIndex: 0,
 };
 
-export const linkStyles: DynamicStyles = {
-	default: {
-		textDecoration: 'none',
-		color: '#2FC5CC',
-		fontSize: '2rem',
-		transition: 'font-size 0.5s ease-in-out',
-	},
-	alternate: {},
-	home: {
-		fontSize: 'clamp(4rem, 6rem, 11vmin)',
-	},
+export const hexagonSvgStyle: CSSObject = {
+	height: '100%',
+	width: '100%',
+	objectFit: 'contain',
+	gridColumn: '1',
+	gridRow: '1',
+};
+
+export const heroTextStyle: CSSObject = {
+	gridColumn: '1',
+	gridRow: '1',
+	zIndex: 2,
+	display: 'flex',
+	flexDirection: 'column',
+	justifyContent: 'center',
+	alignItems: 'center',
+	fontFamily: 'Space Grotesk, var(--font-sans-serif)',
+	alignSelf: 'center',
+	justifySelf: 'center',
+	height: '85%',
+	width: '65%',
 };
 
 export const nameStyle: CSSObject = {
-	lineHeight: '1',
-	margin: '0',
-	boxSizing: 'border-box',
-	paddingBottom: '1rem',
+	fontWeight: 900,
+	fontSize: 'clamp(3rem, 12vmin, 20rem)',
+	margin: '0 0 2rem 0',
+	color: 'transparent',
+	filter: 'drop-shadow(0 0 3px rgba(255, 255, 255, 0.6))',
+	background: 'linear-gradient(180deg, #FFFFFF 0%,rgb(145, 140, 195) 100%)',
+	backgroundClip: 'text',
 };
 
-export const titleStyles: DynamicStyles = {
-	default: {
-		color: '#2FC5CC',
-		fontFamily: 'Courier New',
-		textAlign: 'left',
-	},
-	alternate: {
-		textAlign: 'center',
-	},
-	home: {
-		display: 'flex',
-		flexDirection: 'column',
-		width: '100%',
-	},
+export const titleStyle: CSSObject = {
+	fontWeight: 600,
+	fontSize: 'clamp(1.25rem, 6vmin, 15rem)',
+	color: '#B6ADFF',
+	lineHeight: '1.25',
+	letterSpacing: '0.1em',
+	textAlign: 'center',
+	margin: 0,
+	textShadow: '0 0 3px #B6ADFF',
+};
+
+export const subtitleStyle: CSSObject = {
+	fontWeight: 500,
+	fontSize: 'clamp(1rem, 4vmin, 2rem)',
+	color: '#8575FF',
+	marginTop: '2rem',
+	textShadow: '0 0 2px rgba(148, 134, 255, 0.9)'
 };
